+++ packages/ai/core/index.ts
@@ -1,4 +1,5 @@
-export type { DeepPartial } from '@ai-sdk/ui-utils';
+export { jsonSchema } from '@ai-sdk/ui-utils';
+export type { DeepPartial, Schema } from '@ai-sdk/ui-utils';
 export * from './embed';
 export * from './generate-object';
 export * from './generate-text';
@@ -7,5 +8,3 @@ export * from './registry';
 export * from './tool';
 export * from './types';
 export { cosineSimilarity } from './util/cosine-similarity';
-export { jsonSchema } from './util/schema';
-export type { Schema } from './util/schema';