+++ packages/ai/core/registry/index.ts
@@ -1,4 +1,5 @@
 export * from './invalid-model-id-error';
 export * from './no-such-model-error';
 export * from './no-such-provider-error';
+export type { experimental_Provider } from './provider';
 export * from './provider-registry';