+++ packages/ai/core/registry/invalid-model-id-error.ts
@@ -1,4 +1,12 @@
-export class InvalidModelIdError extends Error {
+import { AISDKError } from '@ai-sdk/provider';
+
+const name = 'AI_InvalidModelIdError';
+const marker = `vercel.ai.error.${name}`;
+const symbol = Symbol.for(marker);
+
+export class InvalidModelIdError extends AISDKError {
+  private readonly [symbol] = true; // used in isInstance
+
   readonly id: string;
 
   constructor({
@@ -8,27 +16,34 @@ export class InvalidModelIdError extends Error {
     id: string;
     message?: string;
   }) {
-    super(message);
-
-    this.name = 'AI_InvalidModelIdError';
+    super({ name, message });
 
     this.id = id;
   }
 
+  static isInstance(error: unknown): error is InvalidModelIdError {
+    return AISDKError.hasMarker(error, marker);
+  }
+
+  /**
+   * @deprecated use `isInstance` instead
+   */
   static isInvalidModelIdError(error: unknown): error is InvalidModelIdError {
     return (
       error instanceof Error &&
-      error.name === 'AI_InvalidModelIdError' &&
+      error.name === name &&
       typeof (error as InvalidModelIdError).id === 'string'
     );
   }
 
+  /**
+   * @deprecated Do not use this method. It will be removed in the next major version.
+   */
   toJSON() {
     return {
       name: this.name,
       message: this.message,
       stack: this.stack,
-
       id: this.id,
     };
   }