+++ packages/ai/rsc/rsc-server.ts
@@ -1,5 +1,5 @@
 export { getAIState, getMutableAIState } from './ai-state';
-export { createStreamableUI, createStreamableValue } from './streamable';
-export { render } from './render';
-export { streamUI } from './stream-ui';
 export { createAI } from './provider';
+export { render, streamUI } from './stream-ui';
+export { createStreamableUI } from './streamable-ui/create-streamable-ui';
+export { createStreamableValue } from './streamable-value/create-streamable-value';