+++ packages/ai/streams/stream-data.ts
@@ -1,6 +1,5 @@
 import { JSONValue, formatStreamPart } from '@ai-sdk/ui-utils';
-
-const STREAM_DATA_WARNING_TIME_MS = 15 * 1000;
+import { HANGING_STREAM_WARNING_TIME_MS } from '../util/constants';
 
 /**
  * A stream wrapper to send custom JSON-encoded data back to the client.
@@ -27,7 +26,7 @@ export class StreamData {
             console.warn(
               'The data stream is hanging. Did you forget to close it with `data.close()`?',
             );
-          }, STREAM_DATA_WARNING_TIME_MS);
+          }, HANGING_STREAM_WARNING_TIME_MS);
         }
       },
       pull: controller => {