+++ packages/google-vertex/src/google-vertex-language-model.ts
@@ -14,6 +14,7 @@ import {
   Part,
   SafetySetting,
   VertexAI,
+  Tool as GoogleTool,
 } from '@google-cloud/vertexai';
 import { convertToGoogleVertexContentRequest } from './convert-to-google-vertex-content-request';
 import {
@@ -110,15 +111,20 @@ export class GoogleVertexLanguageModel implements LanguageModelV1 {
 
     switch (type) {
       case 'regular': {
-        return {
-          model: this.config.vertexAI.getGenerativeModel({
-            model: this.modelId,
-            generationConfig,
-            tools: prepareTools(mode),
-            safetySettings: this.settings.safetySettings as
-              | undefined
-              | Array<SafetySetting>,
+        const conf = {
+          model: this.modelId,
+          generationConfig,
+          tools: prepareTools({
+            mode,
+            useSearchGrounding: this.settings.useSearchGrounding ?? false,
           }),
+          safetySettings: this.settings.safetySettings as
+            | undefined
+            | Array<SafetySetting>,
+        };
+
+        return {
+          model: this.config.vertexAI.getGenerativeModel(conf),
           contentRequest: convertToGoogleVertexContentRequest(prompt),
           warnings,
         };
@@ -199,7 +205,7 @@ export class GoogleVertexLanguageModel implements LanguageModelV1 {
     const { model, contentRequest, warnings } = await this.getArgs(options);
     const { stream } = await model.generateContentStream(contentRequest);
 
-    let finishReason: LanguageModelV1FinishReason = 'other';
+    let finishReason: LanguageModelV1FinishReason = 'unknown';
     let usage: { promptTokens: number; completionTokens: number } = {
       promptTokens: Number.NaN,
       completionTokens: Number.NaN,
@@ -224,13 +230,7 @@ export class GoogleVertexLanguageModel implements LanguageModelV1 {
               const candidate = chunk.candidates?.[0];
 
               if (candidate == null) {
-                controller.enqueue({
-                  type: 'error',
-                  error: new NoContentGeneratedError({
-                    message: 'No candidates in chunk.',
-                  }),
-                });
-                return;
+                return; // ignored (this can happen when using grounding)
               }
 
               if (candidate.finishReason != null) {
@@ -297,18 +297,17 @@ export class GoogleVertexLanguageModel implements LanguageModelV1 {
   }
 }
 
-function prepareTools(
+function prepareTools({
+  useSearchGrounding,
+  mode,
+}: {
+  useSearchGrounding: boolean;
   mode: Parameters<LanguageModelV1['doGenerate']>[0]['mode'] & {
     type: 'regular';
-  },
-) {
+  };
+}): GoogleTool[] | undefined {
   // when the tools array is empty, change it to undefined to prevent errors:
   const tools = mode.tools?.length ? mode.tools : undefined;
-
-  if (tools == null) {
-    return undefined;
-  }
-
   const toolChoice = mode.toolChoice;
 
   if (toolChoice?.type === 'none') {
@@ -316,15 +315,25 @@ function prepareTools(
   }
 
   if (toolChoice == null || toolChoice.type === 'auto') {
-    return [
-      {
-        functionDeclarations: tools.map(tool => ({
-          name: tool.name,
-          description: tool.description ?? '',
-          parameters: prepareFunctionDeclarationSchema(tool.parameters),
-        })),
-      },
-    ];
+    const mappedTools: GoogleTool[] = [];
+
+    if (tools != null) {
+      [
+        {
+          functionDeclarations: tools.map(tool => ({
+            name: tool.name,
+            description: tool.description ?? '',
+            parameters: prepareFunctionDeclarationSchema(tool.parameters),
+          })),
+        },
+      ];
+    }
+
+    if (useSearchGrounding) {
+      mappedTools.push({ googleSearchRetrieval: {} });
+    }
+
+    return mappedTools.length > 0 ? mappedTools : undefined;
   }
 
   // forcing tool calls or a specific tool call is not supported by Vertex: