+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,48 @@
 # ai
 
+## 3.3.7
+
+### Patch Changes
+
+- eccbd8e: feat (ai/core): add onChunk callback to streamText
+- Updated dependencies [dd4a0f5]
+  - @ai-sdk/provider@0.0.19
+  - @ai-sdk/provider-utils@1.0.11
+  - @ai-sdk/ui-utils@0.0.31
+  - @ai-sdk/react@0.0.43
+  - @ai-sdk/solid@0.0.34
+  - @ai-sdk/svelte@0.0.36
+  - @ai-sdk/vue@0.0.35
+
+## 3.3.6
+
+### Patch Changes
+
+- e9c891d: feat (ai/react): useObject supports non-Zod schemas
+- 3719e8a: chore (ai/core): provider registry code improvements
+- Updated dependencies [e9c891d]
+- Updated dependencies [4bd27a9]
+- Updated dependencies [845754b]
+  - @ai-sdk/ui-utils@0.0.30
+  - @ai-sdk/react@0.0.42
+  - @ai-sdk/provider-utils@1.0.10
+  - @ai-sdk/provider@0.0.18
+  - @ai-sdk/solid@0.0.33
+  - @ai-sdk/svelte@0.0.35
+  - @ai-sdk/vue@0.0.34
+
+## 3.3.5
+
+### Patch Changes
+
+- 9ada023: feat (ai/core): mask data stream error messages with streamText
+- Updated dependencies [e5b58f3]
+  - @ai-sdk/ui-utils@0.0.29
+  - @ai-sdk/react@0.0.41
+  - @ai-sdk/solid@0.0.32
+  - @ai-sdk/svelte@0.0.34
+  - @ai-sdk/vue@0.0.33
+
 ## 3.3.4
 
 ### Patch Changes