+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/provider
 
+## 0.0.19
+
+### Patch Changes
+
+- dd4a0f5: fix (ai/provider): remove invalid check in isJSONParseError
+
+## 0.0.18
+
+### Patch Changes
+
+- 4bd27a9: chore (ai/provider): refactor type validation
+
 ## 0.0.17
 
 ### Patch Changes