+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ai-sdk/solid
 
+## 0.0.34
+
+### Patch Changes
+
+- @ai-sdk/provider-utils@1.0.11
+- @ai-sdk/ui-utils@0.0.31
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [e9c891d]
+- Updated dependencies [4bd27a9]
+- Updated dependencies [845754b]
+  - @ai-sdk/ui-utils@0.0.30
+  - @ai-sdk/provider-utils@1.0.10
+
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [e5b58f3]
+  - @ai-sdk/ui-utils@0.0.29
+
 ## 0.0.31
 
 ### Patch Changes